refactor: separate typings from code
