+++ projects/movies/src/app/shared/cdk/infinite-scroll/paginate-state.interface.ts
@@ -1,6 +1,6 @@
 import { PaginatedResult } from '../../state/typings';
-import { LoadingState } from '../withLoadingEmissions';
+import { LoadingState } from '../loading/loading-state.interface';
 
-export type InfiniteScrollState<T extends {}> = PaginatedResult<T> & LoadingState<'loading'>;
-export type PaginationOptions = { page: number, totalPages?: number };
-
+export type InfiniteScrollState<T extends {}> = PaginatedResult<T> &
+  LoadingState<'loading'>;
+export type PaginationOptions = { page: number; totalPages?: number };
