+++ projects/movies/src/app/shared/cdk/loading/loading-state.interface.ts
@@ -1,0 +1,4 @@
+export type DefaultLoadingProp = string & 'loading';
+export type LoadingState<K extends string | DefaultLoadingProp> = {
+  [k in K]: boolean;
+};
