+++ projects/movies/src/app/shared/state/movie.state.ts
@@ -4,12 +4,13 @@ import { TMDBMovieModel } from '../../data-access/api/model/movie.model';
 import { patch, RxState, toDictionary } from '@rx-angular/state';
 import { optimizedFetch } from '../utils/optimized-fetch';
 import { getActions } from '../rxa-custom/actions';
-import { LoadingState, withLoadingEmission } from '../cdk/withLoadingEmissions';
+import { withLoadingEmission } from '../cdk/loading/withLoadingEmissions';
 import {
   getMovie,
   getMovieCategory,
 } from '../../data-access/api/resources/movie.resource';
 import { PaginatedResult } from './typings';
+import { LoadingState } from '../cdk/loading/loading-state.interface';
 
 export interface State
   extends LoadingState<'moviesLoading'>,
